Version 0.4
Refactored code, added the Company entity, made a DataService object to consolidate database access.

+++ src/main/java/com/demo/SpringDBwithUI/ProductRepository.java
@@ -1,13 +1,16 @@
 package com.demo.SpringDBwithUI;
 
 import java.util.List;
+import java.util.Optional;
+
 import org.springframework.data.repository.CrudRepository;
 
-public interface ProductRepository extends CrudRepository<Product, Integer> {
+public interface ProductRepository extends CrudRepository<Product, Long> {
 
     //Product findById(int id);
     List<Product> findByName(String name);
     List<Product> findByNameStartsWithIgnoreCase(String name);
     List<Product> findByNameAndWeight(String name, double weight);
+
 
 }
